Started with web dev in go

+++ testing/main.go
@@ -14,8 +14,8 @@ func main() {
 	log.Println("Result of division: ", result)
 }
 
-func divide(x, y int) (int, error) {
-	var result int
+func divide(x, y float32) (float32, error) {
+	var result float32
 
 	if y == 0 {
 		return result, errors.New("cannot divide by zero")
